Create function for deleting items
The function is based on 3 cases:
- case where node has no children
- case where node has 1 child
- case where node has 2 children

The first case is very easy, the second case is harder because you have
to track the previous node, which edge is correlated with the previous
node, which child exists and based on that information connect the left
or right child with the parent through the left or right edge.

The last case is the hardest to implement, where there is 2 children.
You first have to find the next(in the right branch) smallest number
after the number that is selected for deletion(by going right once,
then keep going left until there is no more left, that means you found
the smallest number), store the data from it, delete it, and "delete"
the first number by replacing it's data with the stored number.